+++ packages/opentelemetry-instrumentation-bedrock/tests/test_meta.py
@@ -1,5 +1,5 @@
 import pytest
-from opentelemetry.semconv.ai import SpanAttributes
+from opentelemetry.semconv_ai import SpanAttributes
 
 import json