+++ packages/opentelemetry-instrumentation-haystack/opentelemetry/instrumentation/haystack/wrap_node.py
@@ -5,7 +5,7 @@ from opentelemetry.instrumentation.utils import (
     _SUPPRESS_INSTRUMENTATION_KEY,
 )
 from opentelemetry.instrumentation.haystack.utils import with_tracer_wrapper
-from opentelemetry.semconv.ai import SpanAttributes, TraceloopSpanKindValues
+from opentelemetry.semconv_ai import SpanAttributes, TraceloopSpanKindValues
 
 logger = logging.getLogger(__name__)