+++ packages/opentelemetry-instrumentation-openai/tests/traces/test_chat.py
@@ -1,5 +1,5 @@
 import pytest
-from opentelemetry.semconv.ai import SpanAttributes
+from opentelemetry.semconv_ai import SpanAttributes
 from openai.types.chat.chat_completion_message_tool_call import (
     ChatCompletionMessageToolCall,
     Function,
@@ -147,6 +147,7 @@ def test_chat_pydantic_based_tool_calls(exporter, openai_client):
 
 @pytest.mark.vcr
 def test_chat_streaming(exporter, openai_client):
+
     response = openai_client.chat.completions.create(
         model="gpt-3.5-turbo",
         messages=[{"role": "user", "content": "Tell me a joke about opentelemetry"}],