+++ packages/opentelemetry-instrumentation-vertexai/tests/disabled_test_bison.py
@@ -2,7 +2,7 @@ import asyncio
 
 import pytest
 import vertexai
-from opentelemetry.semconv.ai import SpanAttributes
+from opentelemetry.semconv_ai import SpanAttributes
 from vertexai.language_models import TextGenerationModel, ChatModel, InputOutputTextPair
 
 vertexai.init()
@@ -30,7 +30,9 @@ def test_vertexai_predict(exporter):
     ]
 
     vertexai_span = spans[0]
-    assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "text-bison@001"
+    assert (
+        vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "text-bison@001"
+    )
     assert (
         "Give me ten interview questions for the role of program manager."
         in vertexai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.user"]
@@ -38,7 +40,10 @@ def test_vertexai_predict(exporter):
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.8
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == response
+    assert (
+        vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
+        == response
+    )
 
 
 @pytest.mark.vcr
@@ -68,7 +73,9 @@ def test_vertexai_predict_async(exporter):
     ]
 
     vertexai_span = spans[0]
-    assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "text-bison@001"
+    assert (
+        vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "text-bison@001"
+    )
     assert (
         "Give me ten interview questions for the role of program manager."
         in vertexai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.user"]
@@ -76,7 +83,10 @@ def test_vertexai_predict_async(exporter):
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.8
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == response
+    assert (
+        vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
+        == response
+    )
 
 
 @pytest.mark.vcr
@@ -109,7 +119,9 @@ def test_vertexai_stream(exporter):
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.8
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == "".join(response)
+    assert vertexai_span.attributes[
+        f"{SpanAttributes.LLM_COMPLETIONS}.0.content"
+    ] == "".join(response)
 
 
 @pytest.mark.vcr
@@ -147,7 +159,9 @@ def test_vertexai_stream_async(exporter):
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.8
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == "".join(response)
+    assert vertexai_span.attributes[
+        f"{SpanAttributes.LLM_COMPLETIONS}.0.content"
+    ] == "".join(response)
 
 
 @pytest.mark.vcr
@@ -182,7 +196,9 @@ def test_vertexai_chat(exporter):
     ]
 
     vertexai_span = spans[0]
-    assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "chat-bison@001"
+    assert (
+        vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "chat-bison@001"
+    )
     assert (
         "How many planets are there in the solar system?"
         in vertexai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.user"]
@@ -190,7 +206,10 @@ def test_vertexai_chat(exporter):
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.95
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == response
+    assert (
+        vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
+        == response
+    )
 
 
 @pytest.mark.vcr
@@ -227,9 +246,13 @@ def test_vertexai_chat_stream(exporter):
     ]
 
     vertexai_span = spans[0]
-    assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "chat-bison@001"
+    assert (
+        vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "chat-bison@001"
+    )
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TOP_P] == 0.95
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_TEMPERATURE] == 0.8
     assert vertexai_span.attributes[SpanAttributes.LLM_REQUEST_MAX_TOKENS] == 256
     assert vertexai_span.attributes[SpanAttributes.LLM_TOP_K] == 40
-    assert vertexai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"] == "".join(response)
+    assert vertexai_span.attributes[
+        f"{SpanAttributes.LLM_COMPLETIONS}.0.content"
+    ] == "".join(response)